+++ src/ConfigProvider.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace SlmQueueBeanstalkd;
+
+class ConfigProvider
+{
+    public function __invoke(): array
+    {
+        $module = new Module();
+        $config = $module->getConfig();
+
+        return [
+            'dependencies'  => $config['service_manager'],
+            'slm_queue'     => $config['slm_queue'],
+        ];
+    }
+}