+++ src/Exeption/ExceptionInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace SlmQueueBeanstalkd\Exception;
+
+use SlmQueue\Exception\ExceptionInterface as SlmQueueExceptionInterface;
+
+/**
+ * Exception for SlmQueueBeanstalkd component
+ */
+interface ExceptionInterface extends SlmQueueExceptionInterface
+{
+}