+++ src/Exeption/LogicException.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace SlmQueueBeanstalkd\Exception;
+
+use LogicException as BaseLogicException;
+
+/**
+ * Logic exception
+ */
+class LogicException extends BaseLogicException implements ExceptionInterface
+{
+}