+++ src/Exeption/RuntimeException.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace SlmQueueBeanstalkd\Exception;
+
+use RuntimeException as BaseRuntimeException;
+
+/**
+ * RuntimeException
+ */
+class RuntimeException extends BaseRuntimeException implements ExceptionInterface
+{
+}