+++ src/Job/JobInterface.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace SlmQueueBeanstalkd\Job;
+
+use Pheanstalk\Contract\JobIdInterface;
+use SlmQueue\Job\JobInterface as SlmJobInterface;
+
+interface JobInterface extends JobIdInterface, SlmJobInterface
+{
+
+}